sis/emo-v02/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
konlpy
JPype1
requests
beautifulsoup4
transformers
torch
jinja2